Move into raw, file_backing and index module
core/anndata.py grew too big and needs to be split up
