+++ anndata/readwrite/h5ad.py
@@ -13,7 +13,8 @@ from pandas.api.types import is_categorical_dtype
 from scipy import sparse
 
 from ..core.sparse_dataset import SparseDataset
-from ..core.anndata import AnnData, Raw
+from ..core.anndata import AnnData
+from ..core.raw import Raw
 from ..compat import _from_fixed_length_strings, _clean_uns, Literal
 from .utils import (
     report_read_key_on_error,
